refactor: move clock module to package root
Eliminates client-to-server layer violation introduced by the
`RawMonotonicClock` change. Re-exports in `server/clock.py`
keep existing test/script imports working.
